# Let Java copy arrays which it can sometimes do faster than manual loops.